move fetch related codes to worker.
