+++ common/src/main/java/org/apache/celeborn/common/memory/MemoryManager.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.network.server.memory;
+package org.apache.celeborn.common.memory;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
