+++ common/src/main/java/org/apache/celeborn/common/memory/ReadBufferDispatcher.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.network.server.memory;
+package org.apache.celeborn.common.memory;
 
 import java.util.ArrayList;
 import java.util.List;
