+++ common/src/main/java/org/apache/celeborn/common/memory/ReadBufferListener.java
@@ -15,29 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.network;
+package org.apache.celeborn.common.memory;
 
-import java.net.InetAddress;
-import java.util.concurrent.Callable;
+import java.util.List;
 
-public class TestUtils {
-  public static String getLocalHost() {
-    try {
-      return InetAddress.getLocalHost().getHostAddress();
-    } catch (Exception e) {
-      throw new RuntimeException(e);
-    }
-  }
+import io.netty.buffer.ByteBuf;
 
-  public static void timeOutOrMeetCondition(Callable<Boolean> callable) throws Exception {
-    int timeout = 10000; // 10s
-    while (true) {
-      if (callable.call() || timeout < 0) {
-        break;
-      }
-
-      timeout = timeout - 100;
-      Thread.sleep(100);
-    }
-  }
+// Do not execute blocking task here.
+public interface ReadBufferListener {
+  void notifyBuffers(List<ByteBuf> allocatedBuffers, Throwable throwable);
 }
