+++ common/src/main/java/org/apache/celeborn/common/memory/ReadBufferRequest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.network.server.memory;
+package org.apache.celeborn.common.memory;
 
 import static com.google.common.base.Preconditions.checkArgument;
