+++ common/src/main/java/org/apache/celeborn/common/memory/WrappedDataBuffer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.network.server.memory;
+package org.apache.celeborn.common.memory;
 
 import io.netty.buffer.ByteBuf;
