+++ common/src/test/java/org/apache/celeborn/common/network/TransportClientFactorySuiteJ.java
@@ -34,6 +34,7 @@ import org.apache.celeborn.common.network.server.BaseMessageHandler;
 import org.apache.celeborn.common.network.server.TransportServer;
 import org.apache.celeborn.common.network.util.TransportConf;
 import org.apache.celeborn.common.util.JavaUtils;
+import org.apache.celeborn.common.util.TestUtils;
 
 public class TransportClientFactorySuiteJ {
   private TransportContext context;
