+++ worker/src/main/scala/org/apache/celeborn/service/deploy/worker/FetchHandler.scala
@@ -34,10 +34,11 @@ import org.apache.celeborn.common.network.buffer.NioManagedBuffer
 import org.apache.celeborn.common.network.client.TransportClient
 import org.apache.celeborn.common.network.protocol._
 import org.apache.celeborn.common.network.protocol.Message.Type
-import org.apache.celeborn.common.network.server.{BaseMessageHandler, ChunkStreamManager}
-import org.apache.celeborn.common.network.server.credit.BufferStreamManager
+import org.apache.celeborn.common.network.server.BaseMessageHandler
 import org.apache.celeborn.common.network.util.{NettyUtils, TransportConf}
 import org.apache.celeborn.common.protocol.PartitionType
+import org.apache.celeborn.service.deploy.worker.fetch.ChunkStreamManager
+import org.apache.celeborn.service.deploy.worker.fetch.credit.BufferStreamManager
 import org.apache.celeborn.service.deploy.worker.storage.{PartitionFilesSorter, StorageManager}
 
 class FetchHandler(val conf: TransportConf) extends BaseMessageHandler with Logging {
