+++ worker/src/main/scala/org/apache/celeborn/service/deploy/worker/fetch/credit/MapDataPartition.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.network.server.credit;
+package org.apache.celeborn.service.deploy.worker.fetch.credit;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -27,13 +27,13 @@ import java.util.concurrent.LinkedBlockingDeque;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.Channel;
+import org.apache.celeborn.common.memory.BufferRecycler;
+import org.apache.celeborn.common.memory.MemoryManager;
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.apache.celeborn.common.meta.FileInfo;
-import org.apache.celeborn.common.network.server.memory.BufferRecycler;
-import org.apache.celeborn.common.network.server.memory.MemoryManager;
 import org.apache.celeborn.common.util.JavaUtils;
 
 public class MapDataPartition {
